requirements.txt
uvicorn==0.22.0
fastapi
boto3==1.37.32
pymongo==4.7.3
python-dotenv
groq==0.22.0
PyPDF2==3.0.1